extract seekbar preferences into it's own patch
